+++ sdk/storage/azure-storage-files-datalake/src/datalake_path_client.cpp
@@ -366,7 +366,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
     return Azure::Response<Models::PathProperties>(std::move(ret), result.ExtractRawResponse());
   }
 
-  Azure::Response<Models::GetPathAccessControlListResult> DataLakePathClient::GetAccessControlList(
+  Azure::Response<Models::PathAccessControlList> DataLakePathClient::GetAccessControlList(
       const GetPathAccessControlListOptions& options,
       const Azure::Core::Context& context) const
   {
@@ -384,9 +384,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
     {
       acl = Models::Acl::DeserializeAcls(result->Acl.GetValue());
     }
-    Models::GetPathAccessControlListResult ret;
-    ret.ETag = std::move(result->ETag);
-    ret.LastModified = std::move(result->LastModified);
+    Models::PathAccessControlList ret;
     if (!acl.HasValue())
     {
       throw Azure::Core::RequestFailedException(
@@ -405,7 +403,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
     {
       ret.Permissions = result->Permissions.GetValue();
     }
-    return Azure::Response<Models::GetPathAccessControlListResult>(
+    return Azure::Response<Models::PathAccessControlList>(
         std::move(ret), result.ExtractRawResponse());
   }